+++ src/auth/clientCredentialsClient.spec.ts
@@ -3,9 +3,9 @@
  * clientCredentials client test
  * @ignore
  */
+import * as assert from 'assert';
 import { BAD_REQUEST, INTERNAL_SERVER_ERROR, OK } from 'http-status';
 import * as nock from 'nock';
-import * as assert from 'power-assert';
 import * as client from '../index';
 
 const DOMAIN = 'DOMAIN';